Remove all warnings on Mac

+++ JASP-Desktop/importers/csv.cpp
@@ -511,7 +511,7 @@ bool CSV::utf16to32(uint32_t &out, char *in, int inSize, int& bytesRead, bool bi
 	}
 
 
-	if (*in >= UNI_SUR_HIGH_START && *in <= UNI_SUR_LOW_START)
+	if ((uint32_t)(*in) >= UNI_SUR_HIGH_START && (uint32_t)(*in) <= UNI_SUR_LOW_START)
 	{
 		if (inSize < 4)
 			return false;
